requirements_pi05.txt
# Requirements for π0.5 Experiments (OpenPI + LIBERO)

# Core OpenPI dependencies (Python >= 3.11)
# Note: For complete installation, use uv package manager
# See: https://docs.astral.sh/uv/getting-started/installation/

# JAX/Flax ecosystem
jax[cuda12]==0.5.3
flax==0.10.2
jaxtyping==0.2.36
equinox>=0.11.8
optax>=0.2.0

# PyTorch (for some components)
torch==2.7.1

# Transformers and tokenizers
transformers==4.53.2
sentencepiece>=0.2.0

# OpenPI core dependencies
augmax>=0.3.4
dm-tree>=0.1.8
einops>=0.8.0
ml_collections==1.0.0
orbax-checkpoint==0.11.13
tyro>=0.9.5
filelock>=3.16.1
beartype==0.19.0
treescope>=0.1.7
rich>=14.0.0
polars>=1.30.0

# Simulation frameworks
robosuite
mujoco
gymnasium>=0.28.1

# Vision processing
opencv-python>=4.10.0.84
Pillow>=11.0.0
imageio>=2.36.1

# Utilities
numpy>=1.22.4,<2.0.0
scipy
pyyaml
tqdm
tqdm-loggable>=0.2
h5py
wandb>=0.19.1

# Cloud storage
fsspec[gcs]>=2.024.6.0

# Additional dependencies for LIBERO
# Note: LIBERO should be installed separately via:
# git clone https://github.com/Lifelong-Robot-Learning/LIBERO.git
# cd LIBERO && pip install -e .
